+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/Repositories/TrustSchoolSearchRepository.cs
@@ -12,6 +12,11 @@ public class TrustSchoolSearchRepository(
     IStringFormattingUtilities stringFormattingUtilities)
     : ITrustSchoolSearchRepository
 {
+    private static readonly HashSet<int> AllowedEstablishmentGroupTypeCodes =
+        Enum.GetValues<NameAndCodeEnums.AllowedEstablishmentGroupTypeCodes>()
+            .Select(static code => (int)code)
+            .ToHashSet();
+
     public async Task<(SearchResult[] Results, SearchResultCount NumberOfResults)>
         GetSearchResultsAsync(string? text, int pageSize, int page = 1)
     {
@@ -44,7 +49,7 @@ public class TrustSchoolSearchRepository(
 
         var filteredEstablishments = establishments
             .Where(x => int.TryParse(x.EstablishmentGroupType?.Code, out var code) &&
-                        Enum.IsDefined(typeof(NameAndCodeEnums.AllowedEstablishmentGroupTypeCodes), code))
+                        AllowedEstablishmentGroupTypeCodes.Contains(code))
             .OrderBy(x => x.Name!.StartsWith(text, StringComparison.OrdinalIgnoreCase) ? 0 : 1)
             .ToList();
 
@@ -59,7 +64,7 @@ public class TrustSchoolSearchRepository(
 
         var allResults = establishmentsResults
             .Concat(trustResults)
-            .OrderBy(x => x.Name)
+            .OrderBy(x => x.Name.StartsWith(text, StringComparison.OrdinalIgnoreCase) ? 0 : 1)
             .ToArray();
 
         // Paging (RESTORED)
@@ -81,6 +86,7 @@ public class TrustSchoolSearchRepository(
         {
             return new SearchResult(
                 e.Urn!.ToString(),
+               null ,
                 e.Name!,
                 e.EstablishmentType!.Name!,
                 stringFormattingUtilities.BuildAddressString(
@@ -97,8 +103,9 @@ public class TrustSchoolSearchRepository(
         {
             return new SearchResult(
                 t.GroupUid!.ToString(),
+                t.ReferenceNumber.ToString(),
                 t.Name!,
-                t.Type!.Name!,
+                t.Type.Name,
                 stringFormattingUtilities.BuildAddressString(
                     t.Address!.Street,
                     t.Address.Locality,