+++ tests/DFE.FindInformationAcademiesTrusts.CypressTests/cypress/e2e/regression/navigation/trust/trust-navigation-core.cy.ts
@@ -16,12 +16,12 @@ describe('Trust Navigation Core Tests', () => {
 
             it('Should check that the Overview navigation button takes me to the overview trust details page', () => {
                 // Governance -> Overview
-                cy.visit('/trusts/governance/trust-leadership?uid=5527');
+                cy.visit('/trusts/governance/trust-leadership?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickOverviewServiceNavButton()
                     .checkOverviewServiceNavButtonIsHighlighted()
-                    .checkCurrentURLIsCorrect('/trusts/overview/trust-details?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/overview/trust-details?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent();
                 overviewPage
                     .checkTrustDetailsSubHeaderPresent();
@@ -29,12 +29,12 @@ describe('Trust Navigation Core Tests', () => {
 
             it('Should check that the contacts navigation button takes me to the contacts in DfE page', () => {
                 // Overview -> Contacts
-                cy.visit('/trusts/overview/trust-details?uid=5527');
+                cy.visit('/trusts/overview/trust-details?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickContactsServiceNavButton()
                     .checkContactsServiceNavButtonIsHighlighted()
-                    .checkCurrentURLIsCorrect('/trusts/contacts/in-dfe?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/contacts/in-dfe?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent();
                 trustContactsPage
                     .checkContactsInDfeSubHeaderPresent();
@@ -42,12 +42,12 @@ describe('Trust Navigation Core Tests', () => {
 
             it('Should check that the Academies navigation button takes me to the academies details page', () => {
                 // Contacts -> Academies
-                cy.visit('/trusts/contacts/in-dfe?uid=5527');
+                cy.visit('/trusts/contacts/in-dfe?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickAcademiesServiceNavButton()
                     .checkAcademiesServiceNavButtonIsHighlighted()
-                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/details?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/details?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent();
                 academiesInTrustPage
                     .checkDetailsHeadersPresent()
@@ -56,12 +56,12 @@ describe('Trust Navigation Core Tests', () => {
 
             it('Should check that the Ofsted navigation button takes me to the Ofsted single headline grades page', () => {
                 //  Academies -> Ofsted
-                cy.visit('/trusts/academies/in-trust/details?uid=5527');
+                cy.visit('/trusts/academies/in-trust/details?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickOfstedServiceNavButton()
                     .checkOfstedServiceNavButtonIsHighlighted()
-                    .checkCurrentURLIsCorrect('/trusts/ofsted/overview?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/ofsted/overview?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent();
                 ofstedPage
                     .checkOfstedSHGPageHeaderPresent();
@@ -69,7 +69,7 @@ describe('Trust Navigation Core Tests', () => {
 
             it('Should check that the Finance documents navigation button takes me to the Financial documents financial statements page', () => {
                 //  Ofsted -> Financial documents 
-                cy.visit('/trusts/ofsted/overview?uid=5527');
+                cy.visit('/trusts/ofsted/overview?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickFinancialDocumentsServiceNavButton()
@@ -82,12 +82,12 @@ describe('Trust Navigation Core Tests', () => {
 
             it('Should check that the Governance navigation button takes me to the governance trust leadership page', () => {
                 //  Financial documents -> Governance
-                cy.visit('/trusts/financial-documents/financial-statements?uid=5527');
+                cy.visit('/trusts/financial-documents/financial-statements?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickGovernanceServiceNavButton()
                     .checkGovernanceServiceNavButtonIsHighlighted()
-                    .checkCurrentURLIsCorrect('/trusts/governance/trust-leadership?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/governance/trust-leadership?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent();
                 governancePage
                     .checkTrustLeadershipSubHeaderPresent();
@@ -97,81 +97,81 @@ describe('Trust Navigation Core Tests', () => {
         describe("Should check that the pipeline academies navigation works", () => {
             it('Navigates from In this trust to pipeline academies', () => {
                 // Academies in Trust -> Pipeline Academies
-                cy.visit('/trusts/academies/in-trust/details?uid=16002');
+                cy.visit('/trusts/academies/in-trust/details?uid=16002&referencenumber=tr04032');
 
                 navigation
                     .clickPipelineAcademiesNavButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/pre-decision?uid=16002');
+                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/pre-decision?uid=16002&referencenumber=tr04032');
             });
 
             it('Navigates from In this trust to pipeline academies', () => {
                 // Pipeline Academies -> Academies in Trust
-                cy.visit('/trusts/academies/pipeline/pre-decision?uid=16002');
+                cy.visit('/trusts/academies/pipeline/pre-decision?uid=16002&referencenumber=tr04032');
 
                 navigation
                     .clickAcademiesInThisTrustNavButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/details?uid=16002');
+                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/details?uid=16002&referencenumber=tr04032');
             });
 
             it('Navigates from In this trust to pipeline academies', () => {
                 // Pipeline Academies page nav buttons
-                cy.visit('/trusts/academies/pipeline/pre-decision?uid=16002');
+                cy.visit('/trusts/academies/pipeline/pre-decision?uid=16002&referencenumber=tr04032');
 
                 navigation
                     .clickPipelineAcademiesPostDecisionNavButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/post-decision?uid=16002')
+                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/post-decision?uid=16002&referencenumber=tr04032')
                     .clickPipelineAcademiesFreeSchoolsNavButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/free-schools?uid=16002')
+                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/free-schools?uid=16002&referencenumber=tr04032')
                     .clickPipelineAcademiesPreDecisionNavButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/pre-decision?uid=16002');
+                    .checkCurrentURLIsCorrect('/trusts/academies/pipeline/pre-decision?uid=16002&referencenumber=tr04032');
             });
         });
 
         describe("Should check that the finance documents navigation works", () => {
             it('Navigates from financial statements to management letters', () => {
                 // Financial statements -> management letters
-                cy.visit('/trusts/financial-documents/financial-statements?uid=5527');
+                cy.visit('/trusts/financial-documents/financial-statements?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickFinancialDocsManagementLettersButton()
-                    .checkCurrentURLIsCorrect('/trusts/financial-documents/management-letters?uid=5527');
+                    .checkCurrentURLIsCorrect('/trusts/financial-documents/management-letters?uid=5527&referencenumber=tr04032');
             });
 
             it('Navigates from management letters to internal scrutiny reports', () => {
                 //Management letters -> Internal scrutiny reports
-                cy.visit('/trusts/financial-documents/management-letters?uid=5527');
+                cy.visit('/trusts/financial-documents/management-letters?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickFinancialDocsInternalScrutinyReportsButton()
-                    .checkCurrentURLIsCorrect('/trusts/financial-documents/internal-scrutiny-reports?uid=5527');
+                    .checkCurrentURLIsCorrect('/trusts/financial-documents/internal-scrutiny-reports?uid=5527&referencenumber=tr04032');
             });
 
             it('Navigates from internal scrutiny reports to self assessment checklist', () => {
                 // Internal scrutiny reports -> Self assessment checklist
-                cy.visit('/trusts/financial-documents/internal-scrutiny-reports?uid=5527');
+                cy.visit('/trusts/financial-documents/internal-scrutiny-reports?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickFinancialDocsSelfAssessmentButton()
-                    .checkCurrentURLIsCorrect('/trusts/financial-documents/self-assessment-checklists?uid=5527');
+                    .checkCurrentURLIsCorrect('/trusts/financial-documents/self-assessment-checklists?uid=5527&referencenumber=tr04032');
             });
 
             it('Navigates from self assessment checklist to financial statements', () => {
                 // Self-assessment checklist -> Finance statements
-                cy.visit('/trusts/financial-documents/self-assessment-checklists?uid=5527');
+                cy.visit('/trusts/financial-documents/self-assessment-checklists?uid=5527&referencenumber=tr04032');
 
                 navigation
                     .clickFinancialDocsFinancialStatementsButton()
-                    .checkCurrentURLIsCorrect('/trusts/financial-documents/financial-statements?uid=5527');
+                    .checkCurrentURLIsCorrect('/trusts/financial-documents/financial-statements?uid=5527&referencenumber=tr04032');
             });
         });
 
         describe("Testing the trust academies page sub navigation", () => {
             // details -> pupil numbers        
             it('Should check that the academies Pupil numbers navigation button takes me to the correct page', () => {
-                cy.visit('/trusts/academies/in-trust/details?uid=5527');
+                cy.visit('/trusts/academies/in-trust/details?uid=5527&referencenumber=tr04032');
                 navigation
                     .clickPupilNumbersAcademiesTrustButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/pupil-numbers?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/pupil-numbers?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent()
                     .checkAllAcademiesNavItemsPresent();
                 academiesInTrustPage
@@ -180,10 +180,10 @@ describe('Trust Navigation Core Tests', () => {
 
             // pupil numbers -> free school meals 
             it('Should check that the academies Free school meals navigation button takes me to the correct page', () => {
-                cy.visit('/trusts/academies/in-trust/pupil-numbers?uid=5527');
+                cy.visit('/trusts/academies/in-trust/pupil-numbers?uid=5527&referencenumber=tr04032');
                 navigation
                     .clickFreeSchoolMealsAcademiesTrustButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/free-school-meals?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/free-school-meals?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent()
                     .checkAllAcademiesNavItemsPresent();
                 academiesInTrustPage
@@ -192,10 +192,10 @@ describe('Trust Navigation Core Tests', () => {
 
             // free school meals -> details  
             it('Should check that the academies Details navigation button takes me to the correct page', () => {
-                cy.visit('/trusts/academies/in-trust/free-school-meals?uid=5527');
+                cy.visit('/trusts/academies/in-trust/free-school-meals?uid=5527&referencenumber=tr04032');
                 navigation
                     .clickDetailsAcademiesTrustButton()
-                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/details?uid=5527')
+                    .checkCurrentURLIsCorrect('/trusts/academies/in-trust/details?uid=5527&referencenumber=tr04032')
                     .checkAllServiceNavItemsPresent()
                     .checkAllAcademiesNavItemsPresent();
                 academiesInTrustPage
@@ -204,7 +204,7 @@ describe('Trust Navigation Core Tests', () => {
             });
 
             it('Should check that the academies sub nav items are not present when I am not in the relevant academies page', () => {
-                cy.visit('/trusts/overview/trust-details?uid=5527');
+                cy.visit('/trusts/overview/trust-details?uid=5527&referencenumber=tr04032');
                 navigation
                     .checkAcademiesSubNavNotPresent();
             });